+++ Changes.md
@@ -7,6 +7,7 @@ you spot any mistakes.
 ## HEAD
 
 * Accept MySQL charset (like `UTF8` or `UTF8MB4`) in `charset` option #808
+* Accept pool options in connection string to `mysql.createPool` #811
 * Clone connection config for new pool connections
 * Default `connectTimeout` to 2 minutes
 * Throw on unknown SSL profile name #817